refactor: reorganize load_data module and remove deprecated scripts
