+++ examples/make_space.sh
